fix:check spreadsheet existence before loading
When trying to load a spreadsheet, if a non-existent name was used, the file read value was returned null. This caused a null Spreadsheet class to be instantiated. Now, before loading a file, an existence check is performed. If it exists, a Spreadsheet class is loaded. Otherwise, an exception is thrown.