feat:add monthly register add sub command
You can create a monthly register using the monthly-register add <args> command. In the Spreadsheet class, the LocalDate class has been replaced by the Date class. This was done due to the need to obtain data for hours, minutes and seconds. The parameter received in the Spreadsheet.addMonthlyRegister(LocalDate date) method has been replaced by two integer parameters, month and year. This was done to allow the user to define the month and year of the monthly register they want to create.